ReusableMergedByteBuffers: Fix NullPointerException when constructed with buffers
This adds a test that previously would fail with a NPE, but now works with the provided fix.
The issue is that fields initialized in the extending class will not be initialized before the super constructor has completed (which is what is invoking to add).
This constructor is thus just problematic and was removed, putting the add burden on the `ReuseableMergedByteBuffers`.