Better handling for internal exceptions
This resolves #52 by allowing the `ExceptionUtils.handleException` behavior to be overriden / controlled.
By default the behavior will stay the same.  The ClientCloseListener and ServerCloseListener provide a default `onCloseWithError` implementation which will invoke into the ExceptionHandler before calling the normal `onClose`.  However if desired this default function can be overriden to avoid having the ExceptionHandler be invoked.

By having this be a default function it allows the CloseListener to still be able to be implemented with a lambda