Client: Synchronize addReadBuffer to ensure it's not run concurrently with getRead()
getRead() typically is invoked on the client thread, and when used that way this should not be a concern.
However if you are not on the reader thread there is the potential for it to try and read while a write is being added.  This is a use case we want to continue to support, so this fixes it by locking while the read is being added.