NoThreadSocketExecutor & SingleThreadSocketExecutor to accept NoThreadScheduler instance
This is to allow stats to be collected on the underline scheduler if desired.
In addition SingleThreadSocketExecutor was transitioned to manually constructing a single thread (rather than a pool), to reduce overhead.  It also will do a `join` on the stop to ensure that the stop has completed before returning.
ThreadedSocketExecutor removed unnecssary variable