Don't .slice() ByteBuffers
This stops doing .slice() operations on ByteBuffers, and instead uses only .duplicate().
The motivation for this is that .slice() makes the use of direct .array() access impossible.
It may be ideal to be able to access the array using the set position / limit so that a copy is not required in order to access the read data.